- Reldens - v4.0.0 - Add RoomsMapEditFormSubscriber for shared map data handling in edit forms and integrate rooms-change-points edit form.
