+++ lib/admin/server/plugin.js
@@ -17,6 +17,7 @@ const { ShutdownSubscriber } = require('./subscribers/shutdown-subscriber');
 const { ThemeManagerSubscriber } = require('./subscribers/theme-manager-subscriber');
 const { RoomsEntitySubscriber } = require('./subscribers/rooms-entity-subscriber');
 const { ObjectsEntitySubscriber } = require('./subscribers/objects-entity-subscriber');
+const { RoomsChangePointsEntitySubscriber } = require('./subscribers/rooms-change-points-entity-subscriber');
 const { GeneratorsRoutesSubscriber } = require('./subscribers/generators-routes-subscriber');
 const { TilesetAnalyzerSubscriber } = require('./subscribers/tileset-analyzer-subscriber');
 const { MapsElementsEditorSubscriber } = require('./subscribers/maps-elements-editor-subscriber');
@@ -111,6 +112,7 @@ class AdminPlugin extends PluginInterface
             this.subscribers.skillsImporter = new SkillsImporterSubscriber(adminManager, themeManager);
             this.subscribers.roomsEntity = new RoomsEntitySubscriber(adminManager, this.config);
             this.subscribers.objectsEntity = new ObjectsEntitySubscriber(adminManager);
+            this.subscribers.roomsChangePointsEntity = new RoomsChangePointsEntitySubscriber(adminManager);
             this.subscribers.generatorsRoutes = new GeneratorsRoutesSubscriber(
                 adminManager,
                 themeManager.projectGenerateDataPath,
