+++ lib/admin/server/templates-list.js
@@ -56,7 +56,8 @@ module.exports.TemplatesList = {
             rooms: 'rooms.html'
         },
         editForm: {
-            objects: 'objects.html'
+            objects: 'objects.html',
+            'rooms-change-points': 'rooms-change-points.html'
         }
     }
 };
